+++ src/ext/java/org/opentripplanner/ext/flex/FlexibleTransitLeg.java
@@ -6,6 +6,7 @@ import java.time.ZonedDateTime;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import javax.annotation.Nonnull;
 import org.locationtech.jts.geom.LineString;
 import org.opentripplanner.ext.flex.edgetype.FlexTripEdge;
 import org.opentripplanner.model.BookingInfo;
@@ -13,9 +14,10 @@ import org.opentripplanner.model.PickDrop;
 import org.opentripplanner.model.plan.Leg;
 import org.opentripplanner.model.plan.Place;
 import org.opentripplanner.model.plan.StopArrival;
+import org.opentripplanner.model.plan.TransitLeg;
 import org.opentripplanner.routing.alertpatch.TransitAlert;
-import org.opentripplanner.routing.core.TraverseMode;
 import org.opentripplanner.transit.model.basic.I18NString;
+import org.opentripplanner.transit.model.basic.TransitMode;
 import org.opentripplanner.transit.model.basic.WheelchairAccessibility;
 import org.opentripplanner.transit.model.framework.AbstractTransitEntity;
 import org.opentripplanner.transit.model.network.Route;
@@ -29,7 +31,7 @@ import org.opentripplanner.util.lang.ToStringBuilder;
  * One leg of a trip -- that is, a temporally continuous piece of the journey that takes place on a
  * particular vehicle, which is running on flexible trip, i.e. not using fixed schedule and stops.
  */
-public class FlexibleTransitLeg implements Leg {
+public class FlexibleTransitLeg implements TransitLeg {
 
   private final FlexTripEdge edge;
 
@@ -53,11 +55,6 @@ public class FlexibleTransitLeg implements Leg {
     this.endTime = endTime;
 
     this.generalizedCost = generalizedCost;
-  }
-
-  @Override
-  public boolean isTransitLeg() {
-    return true;
   }
 
   @Override
@@ -86,8 +83,9 @@ public class FlexibleTransitLeg implements Leg {
   }
 
   @Override
-  public TraverseMode getMode() {
-    return TraverseMode.fromTransitMode(getTrip().getMode());
+  @Nonnull
+  public TransitMode getMode() {
+    return getTrip().getMode();
   }
 
   @Override