+++ src/main/java/org/opentripplanner/api/model/ApiItinerary.java
@@ -105,7 +105,7 @@ public class ApiItinerary {
   public boolean tooSloped = false;
 
   /**
-   * If {@link RouteRequest#allowKeepingRentedVehicleAtDestination}
+   * If {@link RouteRequest#allowArrivingInRentalVehicleAtDestination}
    * is set than it is possible to end a trip without dropping off the rented bicycle.
    */
   public boolean arrivedAtDestinationWithRentedBicycle = false;