+++ src/main/java/org/opentripplanner/routing/algorithm/raptoradapter/transit/request/RoutingRequestTransitDataProviderFilter.java
@@ -1,17 +1,21 @@
 package org.opentripplanner.routing.algorithm.raptoradapter.transit.request;
 
+import java.util.ArrayList;
 import java.util.BitSet;
 import java.util.Collection;
+import java.util.List;
 import java.util.Set;
 import org.opentripplanner.model.modes.AllowTransitModeFilter;
 import org.opentripplanner.routing.algorithm.raptoradapter.transit.TripPatternForDate;
 import org.opentripplanner.routing.api.request.RouteRequest;
 import org.opentripplanner.routing.api.request.StreetMode;
 import org.opentripplanner.routing.api.request.preference.WheelchairAccessibilityPreferences;
+import org.opentripplanner.routing.core.RouteMatcher;
 import org.opentripplanner.transit.model.basic.MainAndSubMode;
 import org.opentripplanner.transit.model.basic.WheelchairAccessibility;
 import org.opentripplanner.transit.model.framework.FeedScopedId;
 import org.opentripplanner.transit.model.network.BikeAccess;
+import org.opentripplanner.transit.model.network.Route;
 import org.opentripplanner.transit.model.network.RoutingTripPattern;
 import org.opentripplanner.transit.model.timetable.Trip;
 import org.opentripplanner.transit.model.timetable.TripTimes;
@@ -39,15 +43,15 @@ public class RoutingRequestTransitDataProviderFilter implements TransitDataProvi
     WheelchairAccessibilityPreferences accessibility,
     boolean includePlannedCancellations,
     Collection<MainAndSubMode> allowedTransitModes,
-    Set<FeedScopedId> bannedRoutes,
-    Set<FeedScopedId> bannedTrips
+    Collection<FeedScopedId> bannedRoutes,
+    Collection<FeedScopedId> bannedTrips
   ) {
     this.requireBikesAllowed = requireBikesAllowed;
     this.wheelchairEnabled = wheelchairEnabled;
     this.wheelchairAccessibility = accessibility;
     this.includePlannedCancellations = includePlannedCancellations;
-    this.bannedRoutes = bannedRoutes;
-    this.bannedTrips = bannedTrips;
+    this.bannedRoutes = Set.copyOf(bannedRoutes);
+    this.bannedTrips = Set.copyOf(bannedTrips);
     this.transitModeFilter = AllowTransitModeFilter.of(allowedTransitModes);
   }
 
@@ -56,13 +60,19 @@ public class RoutingRequestTransitDataProviderFilter implements TransitDataProvi
     TransitService transitService
   ) {
     this(
-      request.modes.transferMode == StreetMode.BIKE,
+      request.journey().transfer().mode() == StreetMode.BIKE,
       request.wheelchair(),
       request.preferences().wheelchairAccessibility(),
       request.preferences().transit().includePlannedCancellations(),
-      request.modes.transitModes,
-      request.getBannedRoutes(transitService.getAllRoutes()),
-      request.bannedTrips
+      request.journey().transit().modes(),
+      bannedRoutes(
+        request.journey().transit().bannedAgencies(),
+        request.journey().transit().bannedRoutes(),
+        request.journey().transit().whiteListedAgencies(),
+        request.journey().transit().whiteListedRoutes(),
+        transitService.getAllRoutes()
+      ),
+      request.journey().transit().bannedTrips()
     );
   }
 
@@ -128,6 +138,89 @@ public class RoutingRequestTransitDataProviderFilter implements TransitDataProvi
       return copy;
     }
     return boardingPossible;
+  }
+
+  public static List<FeedScopedId> bannedRoutes(
+    Collection<FeedScopedId> bannedAgenciesCollection,
+    RouteMatcher bannedRoutes,
+    Collection<FeedScopedId> whiteListedAgenciesCollection,
+    RouteMatcher whiteListedRoutes,
+    Collection<Route> routes
+  ) {
+    if (
+      bannedRoutes.isEmpty() &&
+      bannedAgenciesCollection.isEmpty() &&
+      whiteListedRoutes.isEmpty() &&
+      whiteListedAgenciesCollection.isEmpty()
+    ) {
+      return List.of();
+    }
+
+    Set<FeedScopedId> bannedAgencies = Set.copyOf(bannedAgenciesCollection);
+    Set<FeedScopedId> whiteListedAgencies = Set.copyOf(whiteListedAgenciesCollection);
+
+    List<FeedScopedId> ret = new ArrayList<>();
+    for (Route route : routes) {
+      if (
+        routeIsBanned(bannedAgencies, bannedRoutes, whiteListedAgencies, whiteListedRoutes, route)
+      ) {
+        ret.add(route.getId());
+      }
+    }
+    return ret;
+  }
+
+  /**
+   * Checks if the route is banned. Also, if whitelisting is used, the route (or its agency) has to
+   * be whitelisted in order to not count as banned.
+   *
+   * @return True if the route is banned
+   */
+  private static boolean routeIsBanned(
+    Set<FeedScopedId> bannedAgencies,
+    RouteMatcher bannedRoutes,
+    Set<FeedScopedId> whiteListedAgencies,
+    RouteMatcher whiteListedRoutes,
+    Route route
+  ) {
+    /* check if agency is banned for this plan */
+    if (!bannedAgencies.isEmpty()) {
+      if (bannedAgencies.contains(route.getAgency().getId())) {
+        return true;
+      }
+    }
+
+    /* check if route banned for this plan */
+    if (!bannedRoutes.isEmpty()) {
+      if (bannedRoutes.matches(route)) {
+        return true;
+      }
+    }
+
+    boolean whiteListed = false;
+    boolean whiteListInUse = false;
+
+    /* check if agency is whitelisted for this plan */
+    if (!whiteListedAgencies.isEmpty()) {
+      whiteListInUse = true;
+      if (whiteListedAgencies.contains(route.getAgency().getId())) {
+        whiteListed = true;
+      }
+    }
+
+    /* check if route is whitelisted for this plan */
+    if (!whiteListedRoutes.isEmpty()) {
+      whiteListInUse = true;
+      if (whiteListedRoutes.matches(route)) {
+        whiteListed = true;
+      }
+    }
+
+    if (whiteListInUse && !whiteListed) {
+      return true;
+    }
+
+    return false;
   }
 
   private boolean routeIsNotBanned(TripPatternForDate tripPatternForDate) {