+++ src/test/java/org/opentripplanner/transit/raptor/speed_test/SpeedTestRequest.java
@@ -64,7 +64,7 @@ public class SpeedTestRequest {
     request.setFrom(input.fromPlace());
     request.setTo(input.toPlace());
     request.setNumItineraries(opts.numOfItineraries());
-    request.modes = input.modes();
+    request.journey().setModes(input.modes());
 
     pref
       .transit()
@@ -77,7 +77,12 @@ public class SpeedTestRequest {
       request.setSearchWindow(Duration.ZERO);
     }
 
-    addDebugOptions(request, opts);
+    request
+      .journey()
+      .transit()
+      .raptorDebugging()
+      .withStops(opts.debugStops())
+      .withPath(opts.debugPath());
 
     pref
       .system()
@@ -88,12 +93,7 @@ public class SpeedTestRequest {
           RoutingTag.testCaseCategory(input.category())
         )
       );
-
     return request;
-  }
-
-  private static void addDebugOptions(RouteRequest request, SpeedTestCmdLineOpts opts) {
-    request.raptorDebugging.withStops(opts.debugStops()).withPath(opts.debugPath());
   }
 
   private Instant time(int time) {