+++ docs/Changelog.md
@@ -105,6 +105,7 @@ based on merged pull requests. Search GitHub issues and pull requests for smalle
 - Return typed errors from realtime updates, prepare for realtime statistics [#4424](https://github.com/opentripplanner/OpenTripPlanner/pull/4424)
 - Add mapping for RecordedCall.PredictionInaccurate [#4435](https://github.com/opentripplanner/OpenTripPlanner/pull/4435)
 - Routing request refactor - routing preferences [#4426](https://github.com/opentripplanner/OpenTripPlanner/pull/4426)
+- Extract journey request from RouteRequest [#4446](https://github.com/opentripplanner/OpenTripPlanner/pull/4446)
   [](AUTOMATIC_CHANGELOG_PLACEHOLDER_DO_NOT_REMOVE)
 
 ## 2.1.0 (2022-03-17)